Separate PrefixedULID class to another file
