+++ airframe-ulid/src/main/scala/wvlet/airframe/ulid/ULID.scala
@@ -329,7 +329,3 @@ object ULID {
   }
 
 }
-
-case class PrefixedULID(prefix: String, ulid: ULID, delimiter: Char = ':') {
-  override def toString: String = s"${prefix}${delimiter}${ulid}"
-}
