+++ airframe-ulid/src/test/scala/wvlet/airframe/ulid/ULIDTest.scala
@@ -123,10 +123,4 @@ class ULIDTest extends AirSpec with PropertyCheck {
     val u        = ULID.ofMillis(unixTime)
     u.epochMillis shouldBe unixTime
   }
-
-  test("resourceID") {
-    val ulid = ULID.of(ULID.MinTime, 0, 0)
-    PrefixedULID(prefix = "abcde", ulid = ulid).toString shouldBe "abcde:00000000000000000000000000"
-    PrefixedULID(prefix = "abcde", ulid = ulid, delimiter = '/').toString shouldBe "abcde/00000000000000000000000000"
-  }
 }
